Add files via upload
Moved things around to hopefully organize it better

Separated the 3 lua scripts into their own BED files, this is not how the game processes them but It'll make it easier to read and now the main script is less cluttered. WindowsFunctions.bed is the only one that I am 100% confident that the name is accurate, but the others I just gave it one that made sense.

I'm working towards making Decomp2 compile, it's still missing some external functions but almost all of the syntax errors are gone.